Refactor out the functions from the React components
