+++ src/components/SearchFilters/TriggerGroup.tsx
@@ -2,9 +2,9 @@ import TriggerCode from "@moj-bichard7-developers/bichard7-next-data/dist/types/
 import ConditionalRender from "components/ConditionalRender"
 import IndeterminateCheckbox from "components/IndeterminateCheckbox"
 import { Dispatch } from "react"
-import getLongTriggerCode from "services/entities/transformers/getLongTriggerCode"
 import getShortTriggerCode from "services/entities/transformers/getShortTriggerCode"
 import { FilterAction, ReasonCode } from "types/CourtCaseFilter"
+import { allSelected, noneSelected, selectedTrigger, someSelected } from "utils/triggerGroups/handleSelected"
 import TriggerCheckbox from "./TriggerCheckbox"
 import { TriggerGroupList } from "./TriggerGroup.styles"
 
@@ -13,27 +13,6 @@ interface TriggerGroupProps {
   allGroupTriggers: TriggerCode[]
   filteredReasonCodes: ReasonCode[]
   dispatch: Dispatch<FilterAction>
-}
-
-const selectedTrigger = (triggerCode: string, filteredReasonCodes: ReasonCode[]): boolean | undefined =>
-  !!filteredReasonCodes.find((reasonCode) => getLongTriggerCode(reasonCode.value) === triggerCode)
-
-const someSelected = (allGroupTriggers: TriggerCode[], filteredReasonCodes: ReasonCode[]): boolean => {
-  const some = allGroupTriggers.filter((triggerCode) => selectedTrigger(triggerCode, filteredReasonCodes))
-
-  return some.length > 0 && some.length !== allGroupTriggers.length
-}
-
-const noneSelected = (allGroupTriggers: TriggerCode[], filteredReasonCodes: ReasonCode[]): boolean =>
-  allGroupTriggers.filter((triggerCode) => selectedTrigger(triggerCode, filteredReasonCodes)).length === 0
-
-const allSelected = (allGroupTriggers: TriggerCode[], filteredReasonCodes: ReasonCode[]): boolean => {
-  const selected = filteredReasonCodes.map((reasonCode) => getLongTriggerCode(reasonCode.value))
-
-  if (allGroupTriggers.length === selected.length) {
-    return true
-  }
-  return false
 }
 
 const TriggerGroup = ({ name, allGroupTriggers, filteredReasonCodes, dispatch }: TriggerGroupProps): JSX.Element => {
